+++ tests/files/c_source/multiplex.h
@@ -33,7 +33,6 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-#include <unistd.h>
 
 #ifndef EINVAL
 #    define EINVAL 22
@@ -143,7 +142,7 @@ struct multiplex_message1_t {
  *
  * @return Size of packed data, or negative error code.
  */
-ssize_t multiplex_message1_pack(
+int multiplex_message1_pack(
     uint8_t *dst_p,
     const struct multiplex_message1_t *src_p,
     size_t size);
